Struct dmat_components::components::layouts::sheet::SheetProps

pub struct SheetProps<TsideSignal: Signal<Item = SheetSide> = Always<SheetSide>, Tunderlying_viewSignal: Signal<Item = Option<Dom>> = Always<Option<Dom>>, Tsheet_contentSignal: Signal<Item = Option<Dom>> = Always<Option<Dom>>, TexpandedSignal: Signal<Item = bool> = Always<bool>, Tshow_scrimSignal: Signal<Item = bool> = Always<bool>, TOnScrimClick: Fn(Click) = fn(_: Click), TApplyFn: FnOnce(DomBuilder<HtmlElement>) -> DomBuilder<HtmlElement> = fn(_: DomBuilder<HtmlElement>) -> DomBuilder<HtmlElement>> {
    pub side: TsideSignal,
    pub underlying_view: Tunderlying_viewSignal,
    pub sheet_content: Tsheet_contentSignal,
    pub expanded: TexpandedSignal,
    pub show_scrim: Tshow_scrimSignal,
    pub on_scrim_click: TOnScrimClick,
    pub apply: Option<TApplyFn>,
}

Modal sheets are used to display supplementary content anchored to the sides of the screen They can render a scrim over the underlying UI, and must me dismissed to return to the application

https://material.io/components/sheets-side https://material.io/components/sheets-bottom

Fields
side: TsideSignalunderlying_view: Tunderlying_viewSignalsheet_content: Tsheet_contentSignalexpanded: TexpandedSignalshow_scrim: Tshow_scrimSignalon_scrim_click: TOnScrimClickapply: Option<TApplyFn>
Implementations

impl SheetProps

pub fn new() -> Self

impl<TsideSignal: Signal<Item = SheetSide>, Tunderlying_viewSignal: Signal<Item = Option<Dom>>, Tsheet_contentSignal: Signal<Item = Option<Dom>>, TexpandedSignal: Signal<Item = bool>, Tshow_scrimSignal: Signal<Item = bool>, TOnScrimClick: Fn(Click), TApplyFn: FnOnce(DomBuilder<HtmlElement>) -> DomBuilder<HtmlElement>> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

pub fn side( self, v: SheetSide ) -> SheetProps<Always<SheetSide>, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

pub fn side_signal<TsideSignalNew: Signal<Item = SheetSide>>( self, v: TsideSignalNew ) -> SheetProps<TsideSignalNew, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

impl<TsideSignal: Signal<Item = SheetSide>, Tunderlying_viewSignal: Signal<Item = Option<Dom>>, Tsheet_contentSignal: Signal<Item = Option<Dom>>, TexpandedSignal: Signal<Item = bool>, Tshow_scrimSignal: Signal<Item = bool>, TOnScrimClick: Fn(Click), TApplyFn: FnOnce(DomBuilder<HtmlElement>) -> DomBuilder<HtmlElement>> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

pub fn underlying_view( self, v: Option<Dom> ) -> SheetProps<TsideSignal, Always<Option<Dom>>, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

pub fn underlying_view_signal<Tunderlying_viewSignalNew: Signal<Item = Option<Dom>>>( self, v: Tunderlying_viewSignalNew ) -> SheetProps<TsideSignal, Tunderlying_viewSignalNew, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

impl<TsideSignal: Signal<Item = SheetSide>, Tunderlying_viewSignal: Signal<Item = Option<Dom>>, Tsheet_contentSignal: Signal<Item = Option<Dom>>, TexpandedSignal: Signal<Item = bool>, Tshow_scrimSignal: Signal<Item = bool>, TOnScrimClick: Fn(Click), TApplyFn: FnOnce(DomBuilder<HtmlElement>) -> DomBuilder<HtmlElement>> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

pub fn sheet_content( self, v: Option<Dom> ) -> SheetProps<TsideSignal, Tunderlying_viewSignal, Always<Option<Dom>>, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

pub fn sheet_content_signal<Tsheet_contentSignalNew: Signal<Item = Option<Dom>>>( self, v: Tsheet_contentSignalNew ) -> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignalNew, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

impl<TsideSignal: Signal<Item = SheetSide>, Tunderlying_viewSignal: Signal<Item = Option<Dom>>, Tsheet_contentSignal: Signal<Item = Option<Dom>>, TexpandedSignal: Signal<Item = bool>, Tshow_scrimSignal: Signal<Item = bool>, TOnScrimClick: Fn(Click), TApplyFn: FnOnce(DomBuilder<HtmlElement>) -> DomBuilder<HtmlElement>> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

pub fn expanded( self, v: bool ) -> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignal, Always<bool>, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

pub fn expanded_signal<TexpandedSignalNew: Signal<Item = bool>>( self, v: TexpandedSignalNew ) -> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignalNew, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

impl<TsideSignal: Signal<Item = SheetSide>, Tunderlying_viewSignal: Signal<Item = Option<Dom>>, Tsheet_contentSignal: Signal<Item = Option<Dom>>, TexpandedSignal: Signal<Item = bool>, Tshow_scrimSignal: Signal<Item = bool>, TOnScrimClick: Fn(Click), TApplyFn: FnOnce(DomBuilder<HtmlElement>) -> DomBuilder<HtmlElement>> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

pub fn show_scrim( self, v: bool ) -> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignal, Always<bool>, TOnScrimClick, TApplyFn>

pub fn show_scrim_signal<Tshow_scrimSignalNew: Signal<Item = bool>>( self, v: Tshow_scrimSignalNew ) -> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignalNew, TOnScrimClick, TApplyFn>

impl<TsideSignal: Signal<Item = SheetSide>, Tunderlying_viewSignal: Signal<Item = Option<Dom>>, Tsheet_contentSignal: Signal<Item = Option<Dom>>, TexpandedSignal: Signal<Item = bool>, Tshow_scrimSignal: Signal<Item = bool>, TOnScrimClick: Fn(Click), TApplyFn: FnOnce(DomBuilder<HtmlElement>) -> DomBuilder<HtmlElement>> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

pub fn on_scrim_click<TOnScrimClickNew: Fn(Click)>( self, v: TOnScrimClickNew ) -> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignal, TOnScrimClickNew, TApplyFn>

impl<TsideSignal: Signal<Item = SheetSide>, Tunderlying_viewSignal: Signal<Item = Option<Dom>>, Tsheet_contentSignal: Signal<Item = Option<Dom>>, TexpandedSignal: Signal<Item = bool>, Tshow_scrimSignal: Signal<Item = bool>, TOnScrimClick: Fn(Click), TApplyFn: FnOnce(DomBuilder<HtmlElement>) -> DomBuilder<HtmlElement>> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFn>

pub fn apply<TApplyFnNew: FnOnce(DomBuilder<HtmlElement>) -> DomBuilder<HtmlElement>>( self, v: TApplyFnNew ) -> SheetProps<TsideSignal, Tunderlying_viewSignal, Tsheet_contentSignal, TexpandedSignal, Tshow_scrimSignal, TOnScrimClick, TApplyFnNew>